fix: added better mute/unmute and score display
This includes a change to make use of the synth.volume.value param, combined with useEffect, to better update the volume of the synth.

It also includes the addition of a small delay when initially loading the game to give the user a small interval, and allow for the modals to load.